backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
chromadb
httpx
beautifulsoup4
playwright
sentence-transformers
torch
numpy
google-generativeai
groq
